chore: handle dRep expiry before adapot calculation

+++ aggregates/governance-aggr/src/main/java/com/bloxbean/cardano/yaci/store/governanceaggr/processor/DRepStatusProcessor.java
@@ -106,6 +106,9 @@ public class DRepStatusProcessor {
     @EventListener
     @Transactional
     public void handleRollbackEvent(RollbackEvent rollbackEvent) {
+        if (!governanceAggrProperties.isEnabled())
+            return;
+
         int count = dRepStorage.deleteBySlotGreaterThan(rollbackEvent.getRollbackTo().getSlot());
         log.info("Rollback -- {} drep records", count);
     }
